+++ jieba.cpp
@@ -14,6 +14,17 @@ static char** ConvertWords(const std::vector<std::string>& words) {
   return res;
 }
 
+static Word* ConvertWords(const std::vector<cppjieba::Word>& words) {
+  Word* res = (Word*)malloc(sizeof(Word) * (words.size() + 1));
+  for (size_t i = 0; i < words.size(); i++) {
+    res[i].offset = words[i].offset;
+    res[i].len = words[i].word.size();
+  }
+  res[words.size()].offset = 0;
+  res[words.size()].len = 0;
+  return res;
+}
+
 Jieba NewJieba(const char* dict_path, const char* hmm_path, const char* user_dict) {
   return (Jieba)(new cppjieba::Jieba(dict_path, hmm_path, user_dict));
 }
@@ -52,4 +63,16 @@ char** Tag(Jieba x, const char* sentence) {
     words.push_back(result[i].first + "/" + result[i].second);
   }
   return ConvertWords(words);
+}
+
+Word* Tokenize(Jieba x, const char* sentence, TokenizeMode mode, int is_hmm_used) {
+  std::vector<cppjieba::Word> words;
+  switch (mode) {
+    case SearchMode:
+      ((cppjieba::Jieba*)x)->CutForSearch(sentence, words, is_hmm_used);
+      return ConvertWords(words);
+    default:
+      ((cppjieba::Jieba*)x)->Cut(sentence, words, is_hmm_used);
+      return ConvertWords(words);
+  }
 }